+++ elenasrc2/elc/elc.h
@@ -18,7 +18,7 @@
 #include "errors.h"
 
 // --- ELC common constants ---
-#define ELC_REVISION_NUMBER         0x019B
+#define ELC_REVISION_NUMBER         0x019C
 
 // --- ELC default file names ---
 #ifdef _WIN32