+++ elenasrc2/engine/syntaxtree.h
@@ -172,6 +172,8 @@ enum LexicalType
    lxSealedMultiDispatching   = 0x058181,
    lxResending                = 0x058190,   // resending a message, optional arg - message / -1 (if follow-up operation is available)
    lxGenericResending         = 0x058191,   // resending a generic message
+   lxDirectResending          = 0x058192,   // resending a message, optional arg - message / -1 (if follow-up operation is available)
+   lxSDirectResending         = 0x058193,   // resending a message, optional arg - message / -1 (if follow-up operation is available)
    lxConstantInt              = 0x0181A0,   // arg - reference
    lxConstantString           = 0x0181A1,   // arg - reference
    lxConstantList             = 0x0181A2,   // arg - reference